+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Download the correct version of openH264 (2.2.0)
+- Avoid to register the account while activating it.
 - Update SDK to 5.3.80
 
 ## 5.2.6 - 2024-07-22